Move enforcements.within_max_age to helpers.
